+++ src/lib/app-provider/queries/simulateFee.ts
@@ -1,10 +1,9 @@
 import type { Coin } from "@cosmjs/amino";
 import { SigningCosmWasmClient } from "@cosmjs/cosmwasm-stargate";
-import { useWallet } from "@cosmos-kit/react";
 import { useQuery } from "@tanstack/react-query";
 import { gzip } from "node-gzip";
 
-import { useDummyWallet } from "../hooks";
+import { useCurrentChain, useDummyWallet, useRPCEndpoint } from "../hooks";
 import type {
   AccessType,
   Addr,
@@ -30,25 +29,24 @@ export const useSimulateFeeQuery = ({
   onSuccess,
   onError,
 }: SimulateQueryParams) => {
-  const { address, getCosmWasmClient, currentChainName, currentChainRecord } =
-    useWallet();
+  const { address, getSigningCosmWasmClient, chain } = useCurrentChain();
   const { dummyWallet, dummyAddress } = useDummyWallet();
-
+  const rpcEndpoint = useRPCEndpoint();
   const userAddress = isDummyUser ? dummyAddress : address || dummyAddress;
 
   const simulateFn = async (msgs: ComposedMsg[]) => {
     // TODO: revisit this logic
-    if (!currentChainRecord?.preferredEndpoints?.rpc?.[0] || !userAddress) {
-      throw new Error("No RPC endpoint or user address");
+    if (!userAddress) {
+      throw new Error("No user address");
     }
 
     const client =
       dummyWallet && (isDummyUser || !address)
         ? await SigningCosmWasmClient.connectWithSigner(
-            currentChainRecord.preferredEndpoints.rpc[0],
+            rpcEndpoint,
             dummyWallet
           )
-        : await getCosmWasmClient();
+        : await getSigningCosmWasmClient();
 
     if (!client) {
       throw new Error("Fail to get SigningCosmWasmClient");
@@ -58,11 +56,16 @@ export const useSimulateFeeQuery = ({
   };
 
   return useQuery({
-    queryKey: ["simulate", currentChainName, userAddress, messages],
+    queryKey: [
+      "simulate",
+      chain.chain_name,
+      userAddress,
+      messages,
+      rpcEndpoint,
+    ],
     queryFn: async ({ queryKey }) => simulateFn(queryKey[3] as ComposedMsg[]),
     enabled,
-    keepPreviousData: true,
-    retry: false,
+    retry: 2,
     refetchOnReconnect: false,
     refetchOnWindowFocus: false,
     onSuccess,
@@ -87,13 +90,12 @@ export const useSimulateFeeForStoreCode = ({
   onSuccess,
   onError,
 }: SimulateQueryParamsForStoreCode) => {
-  const { address, getCosmWasmClient, currentChainName } = useWallet();
-
+  const { address, getSigningCosmWasmClient, chain } = useCurrentChain();
   const simulateFn = async () => {
     if (!address) throw new Error("Please check your wallet connection.");
     if (!wasmFile) throw new Error("Fail to get Wasm file");
 
-    const client = await getCosmWasmClient();
+    const client = await getSigningCosmWasmClient();
     if (!client) throw new Error("Fail to get client");
 
     const submitStoreCodeMsg = async () => {
@@ -110,14 +112,14 @@ export const useSimulateFeeForStoreCode = ({
   return useQuery({
     queryKey: [
       "simulate_fee_store_code",
-      currentChainName,
+      chain.chain_name,
       wasmFile,
       permission,
       addresses,
     ],
     queryFn: async () => simulateFn(),
     enabled,
-    retry: false,
+    retry: 2,
     refetchOnReconnect: false,
     refetchOnWindowFocus: false,
     onSuccess,
@@ -160,13 +162,12 @@ export const useSimulateFeeForProposalStoreCode = ({
   onSuccess,
   onError,
 }: SimulateQueryParamsForProposalStoreCode) => {
-  const { address, getCosmWasmClient, currentChainName } = useWallet();
-
+  const { address, getSigningCosmWasmClient, chain } = useCurrentChain();
   const simulateFn = async () => {
     if (!address) throw new Error("Please check your wallet connection.");
     if (!wasmFile) throw new Error("Fail to get Wasm file");
 
-    const client = await getCosmWasmClient();
+    const client = await getSigningCosmWasmClient();
     if (!client) throw new Error("Fail to get client");
 
     const submitStoreCodeProposalMsg = async () => {
@@ -194,7 +195,7 @@ export const useSimulateFeeForProposalStoreCode = ({
   return useQuery({
     queryKey: [
       "simulate_fee_store_code_proposal",
-      currentChainName,
+      chain.chain_name,
       runAs,
       initialDeposit,
       unpinCode,
@@ -208,7 +209,7 @@ export const useSimulateFeeForProposalStoreCode = ({
     ],
     queryFn: async () => simulateFn(),
     enabled,
-    retry: false,
+    retry: 2,
     refetchOnReconnect: false,
     refetchOnWindowFocus: false,
     onSuccess,