+++ src/lib/utils/validate.test.ts
@@ -1,14 +1,20 @@
+/* eslint-disable sonarjs/no-duplicate-string */
 import { isCodeId, isTxHash, isBlock } from "./validate";
 
 describe("isCodeId", () => {
   test("valid", () => {
     expect(isCodeId("1234")).toBeTruthy();
   });
-  test("invalid number string", () => {
-    expect(isCodeId("123E")).toBeFalsy();
-  });
-  test("invalid length", () => {
-    expect(isCodeId("12345678")).toBeFalsy();
+  describe("invalid", () => {
+    test("empty string", () => {
+      expect(isCodeId("")).toBeFalsy();
+    });
+    test("non-number string", () => {
+      expect(isCodeId("123E")).toBeFalsy();
+    });
+    test("length too long", () => {
+      expect(isCodeId("12345678")).toBeFalsy();
+    });
   });
 });
 
@@ -20,15 +26,20 @@ describe("isTxHash", () => {
       )
     ).toBeTruthy();
   });
-  test("invalid hex", () => {
-    expect(
-      isTxHash(
-        "GE7C3F5D7C223257877BCF145452D8B5D45BFD55620B14EE97E6C86EDE16BFE7"
-      )
-    ).toBeFalsy();
-  });
-  test("invalid length", () => {
-    expect(isTxHash("DE7C3F5D7C223257877BCF145452D8B5")).toBeFalsy();
+  describe("invalid", () => {
+    test("empty string", () => {
+      expect(isTxHash("")).toBeFalsy();
+    });
+    test("non-hexstring", () => {
+      expect(
+        isTxHash(
+          "GE7C3F5D7C223257877BCF145452D8B5D45BFD55620B14EE97E6C86EDE16BFE7"
+        )
+      ).toBeFalsy();
+    });
+    test("incorrect length", () => {
+      expect(isTxHash("DE7C3F5D7C223257877BCF145452D8B5")).toBeFalsy();
+    });
   });
 });
 
@@ -36,7 +47,12 @@ describe("isBlock", () => {
   test("valid", () => {
     expect(isBlock("12345678")).toBeTruthy();
   });
-  test("invalid number string", () => {
-    expect(isCodeId("1234ABCD")).toBeFalsy();
+  describe("invalid", () => {
+    test("empty string", () => {
+      expect(isCodeId("")).toBeFalsy();
+    });
+    test("non-number string", () => {
+      expect(isCodeId("1234ABCD")).toBeFalsy();
+    });
   });
 });